Loadable study - Planning Type Test Dode 1. not required in controller 2. change in repo for filter by type Id type: task

+++ api/cp-dss-api/loadable-study/src/main/java/com/cpdss/loadablestudy/entity/LoadableStudy.java
@@ -129,5 +129,5 @@ public class LoadableStudy extends EntityDoc {
   private String sequenceNo;*/
 
   @Column(name = "planning_type_xid")
-  private Long planningTypeXId;
+  private Integer planningTypeXId;
 }
